Изменение типа данных
Изменён тип данных у количества продукта на складе и витрине с unsignedBigInteger на decimal(5,2)

+++ database/migrations/2020_04_05_071214_create_warehouses_table.php
@@ -16,7 +16,7 @@ class CreateWarehousesTable extends Migration
         Schema::create('warehouses', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->unsignedBigInteger('product_id');
-            $table->unsignedBigInteger('amount_of_product');
+            $table->decimal('amount_of_product', 5, 2);
             $table->timestamps();
 
             $table->index('product_id');
